requirements.txt
asgiref==3.5.2
Django==4.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
psycopg2==2.9.3
PyJWT==1.7.1
pytz==2022.2.1
sqlparse==0.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
